requirements.txt
Pillow
zxing-cpp
numpy
Flask
opencv-python
gunicorn
